Fix never ending test on numberFormat
The algorithm tries to find a "+" sign when sometimes there is none or a
"-" sign in present in the expression.
Changing the existing conditions and adding another else case handling
that fixes the issue.
Note that the code is kind of dirty and may not handle all the edge cases.